Update localBench with stats and move more stats collection stuff into infbench package
