+++ inference/python/infbench/resnet50.py
@@ -71,8 +71,8 @@ class resnet50(model.tvmModel, resnet50Base):
     @staticmethod
     def getMlPerfCfg(gpuType, benchConfig):
         if gpuType == "Tesla K20c":
-            maxQps = 28
-            medianLatency = 0.07
+            maxQps = 18
+            medianLatency = 0.065
         elif gpuType == "Tesla V100-SXM2-16GB":
             # Really wish I understood why this was so bad...
             maxQps = 6
@@ -92,8 +92,8 @@ class resnet50Kaas(model.kaasModel, resnet50Base):
     @staticmethod
     def getMlPerfCfg(gpuType, benchConfig):
         if gpuType == "Tesla K20c":
-            maxQps = 31
-            medianLatency = 0.07
+            maxQps = 18
+            medianLatency = 0.070
         elif gpuType == "Tesla V100-SXM2-16GB":
             maxQps = 24
             medianLatency = 0.05
